or_rule

+++ src/domain/main/StoreModule/PurchaseRules/BasketRule.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, String
+from sqlalchemy import Column, Float, String, Integer
 
 from domain.main.Utils import Base_db
 from src.domain.main.StoreModule.PurchaseRules.IRule import IRule
@@ -10,7 +10,7 @@ from src.domain.main.Utils.Response import Response
 class BasketRule(IRule, Base_db.Base):
     __tablename__ = 'basket_rules'
     __table_args__ = {'extend_existing': True}
-    id = Column("id", Float, primary_key=True)
+    id = Column("id", Integer, primary_key=True)
     store_name = Column("store_name", String, primary_key=True)
     min_price = Column("min_price", Float, primary_key=True)
